separate into base table class
